+++ examples/write_to_file.rs
@@ -1,5 +1,5 @@
 fn main() {
-    let x = (0..50).map(|x| (x as f32 / 50.0) * 10.0);
+    let x = (0..50).map(|x| (x as f64 / 50.0) * 10.0);
 
     let mut s = poloto::plot("test", "x", "y");
